+++ src/Json/Store.cpp
@@ -104,7 +104,7 @@ private:
 bool Store::load()
 {
 	auto& root = typeinfo();
-	memcpy_P(rootData.get(), root.defaultData, root.structSize);
+	memcpy_P(getRootData(), root.defaultData, root.structSize);
 	stringPool.clear();
 	arrayPool.clear();
