+++ src/include/ConfigDB/Pool.h
@@ -165,6 +165,24 @@ public:
 };
 
 /**
+ * @brief Identifies array storage within array pool
+ * @note Can't just use uint16_t as it may be unaligned.
+ * Using `alignas` doesn't help.
+ */
+struct __attribute__((packed)) ArrayId {
+	uint8_t value_[2];
+
+	constexpr ArrayId(uint16_t value = 0) : value_{uint8_t(value), uint8_t(value >> 8)}
+	{
+	}
+
+	constexpr operator uint16_t() const
+	{
+		return value_[0] | (value_[1] << 8);
+	}
+};
+
+/**
  * @brief This pool stores array data
  *
  * An array has fixed-size items.
